requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.2
daphne==4.0.0
decorator==5.1.1
Django==5.0.1
djangorestframework==3.14.0
frozenlist==1.4.1
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
multidict==6.0.5
postgis==1.0.4
psycopg==3.1.18
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytz==2023.4
ratelim==0.1.6
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
twilio==8.12.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
websockets==12.0
yarl==1.9.4
zope.interface==6.1
